--- DesignSystem/src/main/java/com/yourssu/design/compose/base/NoRippleButton.kt
@@ -1,61 +1,0 @@
-package com.yourssu.design.compose.base
-
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.*
-import androidx.compose.material.*
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Shape
-import androidx.compose.ui.semantics.Role
-import androidx.compose.ui.unit.dp
-import com.yourssu.design.compose.YdsTheme
-import com.yourssu.design.compose.states.ButtonColorState
-
-@Composable
-fun NoRippleButton(
-    onClick: () -> Unit,
-    colors: ButtonColorState,
-    modifier: Modifier = Modifier,
-    enabled: Boolean = true,
-    showBorder: Boolean = false,
-    interactionSource: MutableInteractionSource = remember { MutableInteractionSource() },
-    elevation: ButtonElevation? = ButtonDefaults.elevation(),
-    shape: Shape = MaterialTheme.shapes.small,
-    contentPadding: PaddingValues = ButtonDefaults.ContentPadding,
-    content: @Composable RowScope.() -> Unit
-) {
-    val contentColor by colors.contentColor(enabled, interactionSource)
-    Surface(
-        modifier = modifier
-            .noRippleClickable(interactionSource, onClick = onClick),
-        shape = shape,
-        color = colors.backgroundColor(enabled, interactionSource).value,
-        contentColor = contentColor,
-        border = if (showBorder) BorderStroke(
-            YdsTheme.border.normal,
-            contentColor
-        ) else null,
-        elevation = elevation?.elevation(enabled, interactionSource)?.value ?: 0.dp,
-    ) {
-        CompositionLocalProvider(LocalContentAlpha provides contentColor.alpha) {
-            ProvideTextStyle(value = MaterialTheme.typography.button) {
-                Row(
-                    modifier = Modifier
-                        .defaultMinSize(
-                            minWidth = ButtonDefaults.MinWidth,
-                            minHeight = ButtonDefaults.MinHeight
-                        )
-                        .padding(contentPadding),
-                    horizontalArrangement = Arrangement.Center,
-                    verticalAlignment = Alignment.CenterVertically,
-                    content = content
-                )
-            }
-        }
-    }
-}
