--- DesignSystem/src/main/java/com/yourssu/design/compose/foundation/Animation.kt
@@ -1,10 +1,0 @@
-package com.yourssu.design.compose.foundation
-
-import androidx.compose.animation.core.CubicBezierEasing
-
-internal val YdsInAndOutEasing = CubicBezierEasing(0.25f, 0.1f, 0.25f, 1f)
-
-enum class Duration(val millis: Int) {
-    Short(100),
-    Medium(250)
-}
