--- DesignSystem/src/main/java/com/yourssu/design/compose/foundation/YdsBasicColors.kt
@@ -1,209 +1,0 @@
-package com.yourssu.design.compose.foundation
-
-import androidx.compose.ui.graphics.Color
-
-// Logo
-internal val LogoIndigo = Color(0xFF4E5986)
-internal val LogoViolet = Color(0xFF8969AC)
-internal val LogoBlue = Color(0xFF1653DB)
-internal val LogoYellow = Color(0xFFFDD655)
-
-// pointColor
-internal val PointColor050 = Color(0xFFEBE6FB)
-internal val PointColor100 = Color(0xFFCBC1F6)
-internal val PointColor200 = Color(0xFFB2A4EF)
-internal val PointColor300 = Color(0xFFA797F0)
-internal val PointColor400 = Color(0xFF816DEC)
-internal val PointColor500 = Color(0xFF5D4DE8)
-internal val PointColor600 = Color(0xFF3230DC)
-internal val PointColor700 = Color(0xFF1B2CD6)
-internal val PointColor800 = Color(0xFF0025CD)
-internal val PointColor900 = Color(0xFF0020C5)
-
-// WarningRed
-internal val WarningRed050 = Color(0xFFFFECEF)
-internal val WarningRed100 = Color(0xFFFFCED4)
-internal val WarningRed200 = Color(0xFFF99C9D)
-internal val WarningRed300 = Color(0xFFF37476)
-internal val WarningRed400 = Color(0xFFFF5252)
-internal val WarningRed500 = Color(0xFFFF3E36)
-internal val WarningRed600 = Color(0xFFF73536)
-internal val WarningRed700 = Color(0xFFE42930)
-internal val WarningRed800 = Color(0xFFD72028)
-internal val WarningRed900 = Color(0xFFC8111C)
-
-// White & Black
-internal val White000 = Color(0xFFFFFFFF)
-internal val Black000 = Color(0xFF101112)
-internal val RealBlack = Color(0xFF000000)
-
-// Gray
-internal val Gray050 = Color(0xFFF8F9FA)
-internal val Gray100 = Color(0xFFF3F5F7)
-internal val Gray200 = Color(0xFFECEEF0)
-internal val Gray300 = Color(0xFFE2E5E8)
-internal val Gray400 = Color(0xFFD4D8DC)
-internal val Gray500 = Color(0xFFB5B9BD)
-internal val Gray600 = Color(0xFF8E9398)
-internal val Gray700 = Color(0xFF505458)
-internal val Gray800 = Color(0xFF3A3D40)
-internal val Gray900 = Color(0xFF252729)
-
-// WB Alpha
-internal val White000A5 = Color(0x0DFFFFFF)
-internal val White000A10 = Color(0x1AFFFFFF)
-internal val White000A30 = Color(0x4DFFFFFF)
-internal val White000A70 = Color(0xB3FFFFFF)
-internal val Black000A5 = Color(0x0D101112)
-internal val Black000A10 = Color(0x1A101112)
-internal val Black000A30 = Color(0x4D101112)
-internal val Black000A70 = Color(0xB3101112)
-internal val Gray900A5 = Color(0x0D252729)
-internal val Gray900A30 = Color(0x4D252729)
-internal val Gray900A70 = Color(0xB3252729)
-
-// Point Alpha
-internal val PointColorA20 = Color(0x33816DEC)
-
-// Pink
-internal val Pink050 = Color(0xFFFDE6F4)
-internal val Pink100 = Color(0xFFFBC1E5)
-internal val Pink200 = Color(0xFFFC98D3)
-internal val Pink300 = Color(0xFFFF6CBF)
-internal val Pink400 = Color(0xFFFF47AD)
-internal val Pink500 = Color(0xFFFF1F99)
-internal val Pink600 = Color(0xFFF21F93)
-internal val Pink700 = Color(0xFFDA1F8B)
-internal val Pink800 = Color(0xFFC31F85)
-
-// Red
-internal val Red050 = Color(0xFFFFEBEE)
-internal val Red100 = Color(0xFFFFCDD2)
-internal val Red200 = Color(0xFFEF9A9A)
-internal val Red300 = Color(0xFFE57373)
-internal val Red400 = Color(0xFFEF5350)
-internal val Red500 = Color(0xFFF44236)
-internal val Red600 = Color(0xFFE53835)
-internal val Red700 = Color(0xFFD32E2F)
-internal val Red800 = Color(0xFFC62728)
-        
-// Orange
-internal val Orange050 = Color(0xFFFFF1E4)
-internal val Orange100 = Color(0xFFFFDBBC)
-internal val Orange200 = Color(0xFFFFC492)
-internal val Orange300 = Color(0xFFFFAC6C)
-internal val Orange400 = Color(0xFFFD9A55)
-internal val Orange500 = Color(0xFFFC8A46)
-internal val Orange600 = Color(0xFFF67F44)
-internal val Orange700 = Color(0xFFEE7141)
-internal val Orange800 = Color(0xFFE6623E)
-
-// Yellow
-internal val Yellow050 = Color(0xFFFFF8DC)
-internal val Yellow100 = Color(0xFFFFF0AB)
-internal val Yellow200 = Color(0xFFFFE672)
-internal val Yellow300 = Color(0xFFFFDC3A)
-internal val Yellow400 = Color(0xFFFFD200)
-internal val Yellow500 = Color(0xFFFBC000)
-internal val Yellow600 = Color(0xFFF6B000)
-internal val Yellow700 = Color(0xFFE1A714)
-internal val Yellow800 = Color(0xFFCC9200)
-
-// Lime
-internal val Lime050 = Color(0xFFF9FBE7)
-internal val Lime100 = Color(0xFFF0F4C3)
-internal val Lime200 = Color(0xFFE5EE9C)
-internal val Lime300 = Color(0xFFDBE775)
-internal val Lime400 = Color(0xFFD3E157)
-internal val Lime500 = Color(0xFFCCDC39)
-internal val Lime600 = Color(0xFFBFCA33)
-internal val Lime700 = Color(0xFFAEB42B)
-internal val Lime800 = Color(0xFF9D9D24)
-
-// Green
-internal val Green050 = Color(0xFFE5F9EE)
-internal val Green100 = Color(0xFFBFF0D5)
-internal val Green200 = Color(0xFF93E7BB)
-internal val Green300 = Color(0xFF5BDE9F)
-internal val Green400 = Color(0xFF0BD588)
-internal val Green500 = Color(0xFF00CC74)
-internal val Green600 = Color(0xFF00BB68)
-internal val Green700 = Color(0xFF00A85B)
-internal val Green800 = Color(0xFF00964F)
-
-// Emerald
-internal val Emerald050 = Color(0xFFE0F8F7)
-internal val Emerald100 = Color(0xFFB3ECEA)
-internal val Emerald200 = Color(0xFF82E0DF)
-internal val Emerald300 = Color(0xFF51D2D4)
-internal val Emerald400 = Color(0xFF2BC9CC)
-internal val Emerald500 = Color(0xFF00BFC7)
-internal val Emerald600 = Color(0xFF00AFB5)
-internal val Emerald700 = Color(0xFF069A9C)
-internal val Emerald800 = Color(0xFF088585)
-
-// Aqua
-internal val Aqua050 = Color(0xFFE0F4FD)
-internal val Aqua100 = Color(0xFFB2E3FA)
-internal val Aqua200 = Color(0xFF7FD0F7)
-internal val Aqua300 = Color(0xFF4CBEF3)
-internal val Aqua400 = Color(0xFF26B0F1)
-internal val Aqua500 = Color(0xFF00A2EE)
-internal val Aqua600 = Color(0xFF0095DF)
-internal val Aqua700 = Color(0xFF0182CB)
-internal val Aqua800 = Color(0xFF0171B7)
-
-// Blue
-internal val Blue050 = Color(0xFFE4F1FA)
-internal val Blue100 = Color(0xFFBEDCF3)
-internal val Blue200 = Color(0xFF98C6EC)
-internal val Blue300 = Color(0xFF74B0E2)
-internal val Blue400 = Color(0xFF489AE7)
-internal val Blue500 = Color(0xFF4B90D7)
-internal val Blue600 = Color(0xFF4582CA)
-internal val Blue700 = Color(0xFF3C71B8)
-internal val Blue800 = Color(0xFF3661A5)
-
-// Indigo
-internal val Indigo050 = Color(0xFFE8EAF6)
-internal val Indigo100 = Color(0xFFC5CAE9)
-internal val Indigo200 = Color(0xFF9FA8DA)
-internal val Indigo300 = Color(0xFF7986CB)
-internal val Indigo400 = Color(0xFF5C6CC0)
-internal val Indigo500 = Color(0xFF3F52B5)
-internal val Indigo600 = Color(0xFF394AAB)
-internal val Indigo700 = Color(0xFF2F409F)
-internal val Indigo800 = Color(0xFF273693)
-
-// Violet
-internal val Violet050 = Color(0xFFEDE7F6)
-internal val Violet100 = Color(0xFFD1C4E9)
-internal val Violet200 = Color(0xFFB39DDB)
-internal val Violet300 = Color(0xFF9575CD)
-internal val Violet400 = Color(0xFF7E57C2)
-internal val Violet500 = Color(0xFF673AB7)
-internal val Violet600 = Color(0xFF5E35B1)
-internal val Violet700 = Color(0xFF512DA8)
-internal val Violet800 = Color(0xFF4527A0)
-
-// Purple
-internal val Purple050 = Color(0xFFF2E5F5)
-internal val Purple100 = Color(0xFFDFBFE7)
-internal val Purple200 = Color(0xFFCA95D8)
-internal val Purple300 = Color(0xFFB56CC8)
-internal val Purple400 = Color(0xFFA54DBB)
-internal val Purple500 = Color(0xFF9632AF)
-internal val Purple600 = Color(0xFF882EA9)
-internal val Purple700 = Color(0xFF7629A0)
-internal val Purple800 = Color(0xFF652597)
-
-// Brown
-internal val Brown050 = Color(0xFFEFEBE9)
-internal val Brown100 = Color(0xFFD7CCC8)
-internal val Brown200 = Color(0xFFBCAAA4)
-internal val Brown300 = Color(0xFFA1887F)
-internal val Brown400 = Color(0xFF8D6E63)
-internal val Brown500 = Color(0xFF795548)
-internal val Brown600 = Color(0xFF6D4C41)
-internal val Brown700 = Color(0xFF5D4037)
-internal val Brown800 = Color(0xFF4E342E)
