--- DesignSystem/src/main/java/com/yourssu/design/compose/foundation/YdsNightBasicColors.kt
@@ -1,39 +1,0 @@
-package com.yourssu.design.compose.foundation
-
-import androidx.compose.ui.graphics.Color
-
-// pointColor
-internal val PointColor050_night = Color(0xFF27233E)
-internal val PointColor100_night = Color(0xFF433A74)
-internal val PointColor200_night = Color(0xFF544895)
-internal val PointColor300_night = Color(0xFF6556B5)
-internal val PointColor400_night = Color(0xFF816DEC)
-internal val PointColor500_night = Color(0xFF9380E8)
-internal val PointColor600_night = Color(0xFFA797F0)
-internal val PointColor700_night = Color(0xFFB2A4EF)
-internal val PointColor800_night = Color(0xFFCBC1F6)
-internal val PointColor900_night = Color(0xFFEBE6FB)
-
-// WarningRed
-internal val WarningRed050_night = Color(0xFF401E1F)
-internal val WarningRed100_night = Color(0xFF5F2627)
-internal val WarningRed200_night = Color(0xFFA93B3B)
-internal val WarningRed300_night = Color(0xFFC74445)
-internal val WarningRed400_night = Color(0xFFFF5252)
-internal val WarningRed500_night = Color(0xFFFF5E5E)
-internal val WarningRed600_night = Color(0xFFF37476)
-internal val WarningRed700_night = Color(0xFFF99C9D)
-internal val WarningRed800_night = Color(0xFFFFCED4)
-internal val WarningRed900_night = Color(0xFFFFECEF)
-
-// Gray
-internal val Gray050_night = Color(0xFF1A1B1C)
-internal val Gray100_night = Color(0xFF1D1E1F)
-internal val Gray200_night = Color(0xFF202122)
-internal val Gray300_night = Color(0xFF2A2B2C)
-internal val Gray400_night = Color(0xFF373839)
-internal val Gray500_night = Color(0xFF5D5E5F)
-internal val Gray600_night = Color(0xFF848586)
-internal val Gray700_night = Color(0xFFBEBFC0)
-internal val Gray800_night = Color(0xFFE1E2E3)
-internal val Gray900_night = Color(0xFFF8F9FA)
