+++ compose/src/main/java/com/yourssu/design/system/compose/foundation/Animation.kt
@@ -1,0 +1,10 @@
+package com.yourssu.design.system.compose.foundation
+
+import androidx.compose.animation.core.CubicBezierEasing
+
+internal val YdsInAndOutEasing = CubicBezierEasing(0.25f, 0.1f, 0.25f, 1f)
+
+enum class Duration(val millis: Int) {
+    Short(100),
+    Medium(250)
+}
