+++ compose/src/main/java/com/yourssu/design/system/compose/states/ButtonColorState.kt
@@ -1,0 +1,43 @@
+package com.yourssu.design.system.compose.states
+
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.interaction.collectIsPressedAsState
+import androidx.compose.runtime.*
+import androidx.compose.ui.graphics.Color
+import com.yourssu.design.system.compose.util.alterColorIfPressed
+
+@Immutable
+data class ButtonColorState(
+    val contentColor: Color = Color.Unspecified,
+    val disabledContentColor: Color = Color.Unspecified,
+    val bgColor: Color = Color.Transparent,
+    val disabledBgColor: Color = Color.Transparent
+) {
+    @Composable
+    fun contentColor(
+        enabled: Boolean,
+        interactionSource: MutableInteractionSource
+    ): State<Color> {
+        val pressed = interactionSource.collectIsPressedAsState().value
+
+        return rememberUpdatedState(if (enabled) {
+            contentColor.alterColorIfPressed(isPressed = pressed)
+        } else {
+            disabledContentColor
+        })
+    }
+
+    @Composable
+    fun backgroundColor(
+        enabled: Boolean,
+        interactionSource: MutableInteractionSource
+    ): State<Color> {
+        val pressed by interactionSource.collectIsPressedAsState()
+
+        return rememberUpdatedState(if (enabled) {
+            bgColor.alterColorIfPressed(isPressed = pressed)
+        } else {
+            disabledBgColor
+        })
+    }
+}
