+++ settings.gradle
@@ -2,3 +2,4 @@ include ':DesignSystem'
 include ':app:storybook'
 include ':app:yds-ui-tester'
 rootProject.name = "YDS-Android"
+include ':compose'
